we use alignword(sizeof(size_t)) in dealloc, and these need to match (int and size_t are different sizes on 64 bit architectures).  Also, it really seems like these should be alignword(2*sizeof(size_t)), no?  There are 2 size_t sized blocks at the head of the block, it seems like if we happen to word-align on the initial allocation, we could correct the second of those.
Can someone who knows this bit better comment n that?

+++ apc_sma.c
@@ -175,7 +175,7 @@ static int sma_allocate(void* shmaddr, size_t size)
     }
     header->nfoffset = last_offset;
 
-    return OFFSET(cur) + alignword(sizeof(int));
+    return OFFSET(cur) + alignword(sizeof(size_t));
 }
 /* }}} */
